microservice/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
Pillow==6.1.0
pkg-resources==0.0.0
pytz==2019.2
sqlparse==0.3.0
